+++ spec/active_record_compose/inner_model_spec.rb
@@ -4,22 +4,24 @@ require 'spec_helper'
 require 'active_record_compose/inner_model_collection'
 
 RSpec.describe ActiveRecordCompose::InnerModel do
+  let(:owner) { nil }
+
   describe '#==' do
-    subject(:inner_model) { ActiveRecordCompose::InnerModel.new(account, context: :save) }
+    subject(:inner_model) { ActiveRecordCompose::InnerModel.new(nil, account, context: :save) }
 
     let(:account) { Account.new }
     let(:profile) { Profile.new }
 
     specify 'returns true if and only if model and context are equivalent' do
       expect(inner_model).not_to eq nil
-      expect(inner_model).not_to eq ActiveRecordCompose::InnerModel.new(profile)
-      expect(inner_model).not_to eq ActiveRecordCompose::InnerModel.new(account, context: :destroy)
-      expect(inner_model).to eq ActiveRecordCompose::InnerModel.new(account, context: :save)
+      expect(inner_model).not_to eq ActiveRecordCompose::InnerModel.new(owner, profile)
+      expect(inner_model).not_to eq ActiveRecordCompose::InnerModel.new(owner, account, context: :destroy)
+      expect(inner_model).to eq ActiveRecordCompose::InnerModel.new(owner, account, context: :save)
     end
   end
 
   describe '#save' do
-    subject(:inner_model) { ActiveRecordCompose::InnerModel.new(already_persisted_account, context:) }
+    subject(:inner_model) { ActiveRecordCompose::InnerModel.new(nil, already_persisted_account, context:) }
 
     let(:already_persisted_account) { Account.create(name: 'foo', email: 'foo@example.com') }